requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
